Login method in LogInController now spawns a new Thread when a user log in. The Thread repeatedly checks if the currentUser is still authenticated.
Get, Post, Put and Delete methods in ServerConnection now call the Logout method if a status code 401 Unauthorized is received in the response.

The Logout method sets all values of the Session to null and the Thread that was spawned when the user logged in will redirect the user back to the LogInPane.

Also made a little progress on the getOpenGames method