Added functionality for showOpenGames() in PlayMenuController. The user can now see all open joinable games, meaning the user isn't the host of these.
Also added functionality for playBtn and joinBtn in PlayMenuController - they now call ServerConnection's methods joinGame(). playBtn now sets a the selected game as a static variable and redirects the user to the PlayGamePane where the user can enter his/hers controls, press playBtn and get the result on a dialog box. When the user close the dialog box it is then redirected back to the PlayMenuPane where it can proceed with joining, playing or creating games.

Also made minor fixes such as renaming.