Added screen for creating games and functionality for it in the CreateGameController class. It is now possible to create games as either open game or invite another player from the table.
Also added delete button in PlayMenuPane and functionality in PlayMenuController. The user can now delete games that are not flagged "finished" if the user is a player in the game.
Right now the user can delete games in which it is the opponent, however this might not be desirable as a user can exploit this to join and delete other players' open games.
The reason a user is able to delete games as an opponent is that the player might have been invited to the game and doesn't want to play it.

Suggestion:
Make it possible for players to leave games instead of deleting them.